agents/secretary_agent/requirements.txt
fastapi
uvicorn
openai-whisper
pydantic
requests
python-multipart